+++ R/plot_functions.R
@@ -924,7 +924,7 @@ venn_pq <-
 #' @return A \code{\link{ggplot}}2 plot representing Venn diagramm of
 #'   modalities of the argument \code{factor} or if split_by is set a list
 #'   of plots.
-#'
+#' @seealso [upset_pq()] 
 #' @examples
 #' data(data_fungi)
 #' ggvenn_pq(data_fungi, fact = "Height")
@@ -2272,7 +2272,7 @@ SRS_curve_pq <- function(physeq, clean_pq = FALSE, ...) {
 }
 
 
-#' Visualization of two samples for comparison
+#' iNterpolation and EXTrapolation of Hill numbers (with iNEXT)
 #' @description
 #' `r lifecycle::badge("experimental")`
 #'
@@ -2281,13 +2281,13 @@ SRS_curve_pq <- function(physeq, clean_pq = FALSE, ...) {
 #'   physeq are merged using the vector set by `merge_sample_by`. This
 #'   merging used the [speedyseq::merge_samples2()]. In the case of
 #'   [biplot_pq()] this must be a factor with two levels only.
-#' @param ... other arguments for the ggplot function
-#' @return A object of class XXX
+#' @param ... other arguments for the [iNEXT::iNEXT()] function
+#' @return see [iNEXT::iNEXT()] documentation
 #' @export
 #'
 #' @examples
 #' library("iNEXT")
-#' iNEXT_pq(data_fungi, merge_sample_by="Height", q=1, datatype="abundance", nboot=5)
+#' res_iNEXT <- iNEXT_pq(data_fungi, merge_sample_by="Height", q=1, datatype="abundance", nboot=5)
 #' ggiNEXT(res_iNEXT)
 #' ggiNEXT(res_iNEXT, type = 2)
 #' ggiNEXT(res_iNEXT, type = 3)
@@ -2306,3 +2306,212 @@ iNEXT_pq <- function(physeq, merge_sample_by = NULL, ...){
 }
 
 
+
+
+#' Make upset plot for phyloseq object.
+#' @description
+#' `r lifecycle::badge("experimental")`
+#' 
+#' Alternative to venn plot.
+#' 
+#' @inheritParams clean_pq
+#' @param fact (required): Name of the factor to cluster samples by modalities.
+#'   Need to be in \code{physeq@sam_data}.
+#' @param min_nb_seq minimum number of sequences by OTUs by
+#'   samples to take into count this OTUs in this sample. For example,
+#'   if min_nb_seq=2,each value of 2 or less in the OTU table
+#'   will not count in the venn diagramm
+#' @param na_remove : if TRUE (the default), NA values in fact are removed
+#'   if FALSE, NA values are set to "NA" 
+#' @param numeric_fonction (default : sum) the function for numeric vector
+#' @param ... other arguments passed on to the [ComplexUpset::upset()]
+#'
+#' @return A \code{\link{ggplot}}2 plot 
+#' @export
+#' @author Adrien Taudière
+#'
+#' @seealso [ggvenn_pq()]
+#' @examples 
+#' upset_pq(data_fungi, modality = "Height", width_ratio = 0.2)
+#' upset_pq(data_fungi, modality = "Height", min_nb_seq = 1000)
+#' upset_pq(data_fungi, modality = "Height", na_remove = FALSE)
+#' upset_pq(data_fungi, modality = "Time", width_ratio = 0.2)
+#' 
+#' upset_pq(
+#'   data_fungi,
+#'   modality = "Time",
+#'   width_ratio = 0.2,
+#'   annotations = list(
+#'     'Sequences per ASV \n (log10)' = (
+#'       ggplot(mapping = aes(y = log10(Abundance)))
+#'       + geom_jitter(aes(color =
+#'                           Abundance), na.rm = TRUE)
+#'       + geom_violin(alpha = 0.5, na.rm = TRUE) +
+#'         theme(legend.key.size = unit(0.2, "cm")) +
+#'         theme(axis.text=element_text(size=12))
+#'     ),
+#'     'ASV per phylum' = (
+#'       ggplot(mapping = aes(fill = Phylum))
+#'       + geom_bar() + ylab("ASV per phylum") +
+#'         theme(legend.key.size = unit(0.2, "cm"))+
+#'         theme(axis.text=element_text(size=12))
+#'     )
+#'   )
+#' )
+#' 
+#' 
+#' upset_pq(
+#'   subset_taxa(data_fungi, Phylum == "Basidiomycota"),
+#'   modality = "Time",
+#'   width_ratio = 0.2,
+#'   base_annotations=list(),
+#'   annotations = list(
+#'     'Sequences per ASV \n (log10)' = (
+#'       ggplot(mapping = aes(y = log10(Abundance)))
+#'       + geom_jitter(aes(color =
+#'                           Abundance), na.rm = TRUE)
+#'       + geom_violin(alpha = 0.5, na.rm = TRUE) +
+#'         theme(legend.key.size = unit(0.2, "cm")) +
+#'         theme(axis.text=element_text(size=12))
+#'     ),
+#'     'ASV per phylum' = (
+#'       ggplot(mapping = aes(fill = Class))
+#'       + geom_bar() + ylab("ASV per Class") +
+#'         theme(legend.key.size = unit(0.2, "cm"))+
+#'         theme(axis.text=element_text(size=12))
+#'     )
+#'   )
+#' )
+#' 
+#' data_fungi2 <- data_fungi
+#' data_fungi2@sam_data[["Time_0"]] <- data_fungi2@sam_data$Time == 0
+#' data_fungi2@sam_data[["Height__Time_0"]] <-
+#'   paste0(data_fungi2@sam_data[["Height"]] , "__", data_fungi2@sam_data[["Time_0"]])
+#' data_fungi2@sam_data[["Height__Time_0"]][grepl("NA", data_fungi2@sam_data[["Height__Time_0"]])] <-
+#'   NA
+#' upset_pq(data_fungi2, modality = "Height__Time_0", width_ratio = 0.2)
+upset_pq <-
+  function(physeq,
+           fact,
+           min_nb_seq = 0,
+           na_remove = TRUE,
+           numeric_fonction = sum,
+           ...) {
+    if (!is.null(min_nb_seq)) {
+      physeq <- subset_taxa_pq(physeq, taxa_sums(physeq) >= min_nb_seq)
+    }
+
+    if (na_remove) {
+      physeq <-
+        subset_samples_pq(physeq,!is.na(physeq@sam_data[[fact]]))
+    } else {
+      physeq@sam_data[[fact]][is.na(physeq@sam_data[[fact]])] <-
+        "NA"
+    }
+
+    physeq <- speedyseq::merge_samples2(physeq, fact)
+
+    psm <- psmelt(physeq)
+    samp_names <- unique(psm$Sample)
+    psm <-
+      psm %>% mutate(val = TRUE) %>% pivot_wider(names_from = Sample, values_from = val)
+    psm[samp_names][is.na(psm[samp_names])] <- FALSE
+
+    psm <- psm %>% filter(Abundance != 0)
+    psm[[fact]] <- as.character(psm[[fact]])
+
+
+    psm2 <- data.frame(lapply(psm, function(col) {
+      tapply(col, paste0(psm$OTU), function(vec) {
+        diff_fct_diff_class(vec, numeric_fonction = numeric_fonction, na.rm = TRUE)
+      })
+    })) %>% arrange(., desc(Abundance))
+
+    colnames(psm2) <- colnames(psm)
+
+    p <-
+      ComplexUpset::upset(psm2, intersect = samp_names, ...) + xlab(fact)
+    return(p)
+  }
+
+
+#' Compute different functions for different class of vector.
+#' 
+#' @description
+#' `r lifecycle::badge("experimental")`
+#' Mainly an internal function useful in "lapply(..., tapply)" methods
+#'
+#' 
+#' @param x : a vector
+#' @param numeric_fonction : a function for numeric vector. For ex. `sum` or `mean`
+#' @param logical_method : A method for logical vector. One of :
+#'   - TRUE_if_one
+#'   - NA_if_not_all_TRUE
+#'   - FALSE_if_not_all_TRUE
+#' @param character_method : A method for character vector (and factor). One of :
+#'   - unique_or_na
+#'   - more frequent
+#'   - more_frequent_without_equality
+#' @param ... other arguments passed on to the numeric function (ex. na.rm=TRUE)
+#' @return a single value
+#' @export
+#' 
+#' @author Adrien Taudière
+diff_fct_diff_class <-
+  function(x,
+           numeric_fonction = mean,
+           na.rm = TRUE,
+           logical_method = "true_if_one",
+           ...) {
+    if (is.character(x) || is.factor(x)) {
+      if (length(unique(x)) == 1) {
+        return(unique(x))
+      }
+      else if(character_method == "unique_or_na") {
+        return(NA)
+      } else if(character_method == "more_frequent") {
+        return(names(sort(table(x), decreasing = TRUE)[1]))
+      } else if(character_method == "more_frequent_without_equality") {
+        if(names(sort(table(x), decreasing = TRUE)[1]) == names(sort(table(x), decreasing = TRUE)[2])) { 
+          return(NA)
+        } else {
+          return(names(sort(table(x), decreasing = TRUE)[1]))
+        }
+      } else {
+        stop(paste0(character_method, " is not a valid method for character_method params."))
+      }
+    } else if (is.numeric(x)) {
+      return(numeric_fonction(x, ...))
+    } else if (is.logical(x)) {
+      if (logical_method == "TRUE_if_one") {
+        if (sum(x, na.rm = na.rm) > 0) {
+          return(TRUE)
+        } else {
+          return(FALSE)
+        }
+      }
+      if (logical_method == "NA_if_not_all_TRUE") {
+        if (sum(x, na.rm = na.rm) > 0 && sum(!x, na.rm = na.rm) == 0) {
+          return(TRUE)
+        }
+        else if (sum(!x, na.rm = na.rm) > 0 && sum(x, na.rm = na.rm) > 0) {
+          return(NA)
+        } else if (sum(!x, na.rm = na.rm) > 0 && sum(x, na.rm = na.rm) == 0) {
+          return(FALSE)
+        }
+      }
+      if (logical_method == "FALSE_if_not_all_TRUE") {
+        if (sum(x, na.rm = na.rm) > 0 && sum(!x, na.rm = na.rm) == 0) {
+          return(TRUE)
+        }
+        else  {
+          return(FALSE)
+        }
+      } else {
+        stop(paste0(character_method, " is not a valid method for character_method params."))
+      }
+    }
+    else {
+      stop("At least one column is neither numeric nor character or logical")
+    }
+  }